server/requirements.txt
pyyaml
sanic
stravalib
